+++ sdk/cosmosdb/Azure.ResourceManager.CosmosDB/CHANGELOG.md
@@ -10,6 +10,15 @@
 
 ### Other Changes
 
+- Migrated from Swagger to TypeSpec-based generation
+
+## 1.4.1 (2026-06-03)
+
+### Other Changes
+
+- Upgraded dependent Azure.Core to 1.57.0.
+- Upgraded dependent Azure.ResourceManager to 1.14.0.
+
 ## 1.4.0 (2025-12-05)
 
 ### Features Added